refactor(responsecache): remove legacy Middleware() path
Production traffic has gone exclusively through HandleRequest (translated
inference service) and HandleInternalRequest (guardrail executor) since the
exchange seam landed; the echo middleware wrapper was test-only. Remove
ResponseCacheMiddleware.Middleware(), simpleCacheMiddleware.Middleware(), and
the three helpers only that path used (requestBodyForCache,
shouldSkipCacheForWorkflow, shouldSkipCache) — their production equivalents
live in handle() or at the call sites, which pass the patched body explicitly
and gate on workflow.CacheEnabled().

Test migration (possible-refactoring #4): middleware_test.go becomes
exact_cache_test.go. Hit/miss, key separation, streaming-vs-JSON replay,
Cache-Control skip, close-drains-writes, and write-concurrency tests now
drive the production HandleRequest entry. Dropped with the legacy path, as
their behavior no longer exists in this package: workflow-gating bypass tests
(caller-gated since the HandleRequest migration), snapshot-body and
body-read-error tests (callers supply the body), and the non-cacheable-path
test (exact-layer path gating was legacy-only; the semantic layer still
enforces cacheablePaths and callers only invoke the cache for inference
endpoints).
